requirements.txt
# Core RAG con LangChain (LCEL incluido)
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-openai>=0.3.0
langchain-text-splitters>=0.3.0
langchain-core>=0.3.0

# Vector database (versión compatible con langchain-chroma)
chromadb>=1.0.20

# Document processing
pypdf>=4.0.0
python-dotenv>=1.0.0

# Generación (OpenAI)
openai>=1.0.0

# Embeddings locales (opcional - requiere Python <=3.12)
# NOTA: Python 3.13 no soporta PyTorch todavía
# Si usas Python 3.13, usa embedding_model="openai" en lugar de sentence-transformers
torch>=2.0.0; python_version < '3.13'
sentence-transformers>=2.0.0; python_version < '3.13'

# Evaluation
ragas>=0.1.9
datasets>=2.14.0

# Utilities (versiones compatibles)
numpy>=1.21.0,<=2.1.0
pandas>=1.3.0
